Minor pokes.

+++ PlayerPROKit/PlayerPROCoreAdditions.swift
@@ -168,8 +168,8 @@ public let equalizerPacketElements = 512
 /// **Normally** it is never called, only when a FATAL error has occured. <br>
 /// This function is automatically invoked using the macros `#line` and
 /// `#file` for the `line` and `file` paramaters, respectively.
-public func MADDebug(string text: String, line: UInt = #line, file: StaticString = #file) {
-	MADDebugStr(Int16(line), (String(describing: file) as NSString).fileSystemRepresentation, text)
+public func MADDebug(string text: String, line: Int16 = #line, file: StaticString = #file) {
+	MADDebugStr(line, (String(describing: file) as NSString).fileSystemRepresentation, text)
 }
 
 private var BlankNameChar32: (Int8, Int8, Int8, Int8, Int8, Int8, Int8, Int8, Int8, Int8, Int8,
